+++ src/FixedNumber.ts
@@ -105,7 +105,8 @@ export class FixedNumber implements ExactNumberType {
     }
 
     const fixedOperand = operand as FixedNumber;
-    const res = new FixedNumber(this.number * fixedOperand.number, this.decimalPos + fixedOperand.decimalPos);
+    const product = this.number * fixedOperand.number;
+    const res = new FixedNumber(product, this.decimalPos + fixedOperand.decimalPos);
     return res;
   }
 
@@ -197,23 +198,33 @@ export class FixedNumber implements ExactNumberType {
     return this.round(decimals, RoundingMode.TO_ZERO);
   }
 
-  round(decimals?: number, roundingMode?: RoundingMode) {
-    decimals = decimals === undefined ? 0 : decimals;
-    if (!Number.isSafeInteger(decimals) || decimals < 0) {
-      throw new Error('Invalid value for decimals');
+  private isTieStr(lastDigitsStr: string) {
+    if (lastDigitsStr[0] !== '5') return false;
+    for (let i = 1; i < lastDigitsStr.length; i++) {
+      if (lastDigitsStr[i] !== '0') {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private _round(decimals?: number, roundingMode?: RoundingMode) {
+    const shift = this.decimalPos - decimals;
+    if (shift <= 0) {
+      return this;
     }
 
-    const shift = this.decimalPos - decimals;
-    const exp = 10n ** BigInt(Math.abs(shift));
-
-    const outDigits = shift > 0 ? this.number / exp : this.number * exp;
+    const exp = 10n ** BigInt(shift);
+    const outDigits = this.number / exp;
 
     if (roundingMode === RoundingMode.TO_ZERO) {
       return new FixedNumber(outDigits, decimals);
     }
 
-    const extraDigits = shift > 0 ? this.number % exp : 0n;
-    if (extraDigits === 0n) return new FixedNumber(outDigits, decimals);
+    const extraDigits = this.number % exp;
+    if (extraDigits === 0n) {
+      return new FixedNumber(outDigits, decimals);
+    }
 
     if (roundingMode === RoundingMode.AWAY_FROM_ZERO) {
       const res = this.number < 0n ? outDigits - 1n : outDigits + 1n;
@@ -246,22 +257,11 @@ export class FixedNumber implements ExactNumberType {
     let extraDigitsStr = (extraDigits < 0n ? -extraDigits : extraDigits).toString();
     // '00123' extra part will appear in extraDigitsStr as '123'
     // -> in this case we can exclude the tie case by setting the extra part to zero
-    const expectedFracDecimals = shift > 0 ? shift : 0;
-    if (extraDigitsStr.length < expectedFracDecimals) {
+    if (extraDigitsStr.length < shift) {
       extraDigitsStr = '0';
     }
 
-    let isTie = extraDigitsStr[0] === '5';
-    if (isTie) {
-      for (let i = 1; i < extraDigitsStr.length; i++) {
-        if (extraDigitsStr[i] !== '0') {
-          isTie = false;
-          break;
-        }
-      }
-    }
-
-    if (isTie) {
+    if (this.isTieStr(extraDigitsStr)) {
       if (roundingMode === RoundingMode.NEAREST_TO_ZERO) {
         return new FixedNumber(outDigits, decimals);
       }
@@ -299,6 +299,15 @@ export class FixedNumber implements ExactNumberType {
     return new FixedNumber(res, decimals);
   }
 
+  round(decimals?: number, roundingMode?: RoundingMode) {
+    decimals = decimals === undefined ? 0 : decimals;
+    if (!Number.isSafeInteger(decimals) || decimals < 0) {
+      throw new Error('Invalid value for decimals');
+    }
+
+    return this._round(decimals, roundingMode).normalize();
+  }
+
   _incExponent(amount: number): FixedNumber {
     if (amount === 0) return this;
     let newNumber = this.number;
@@ -319,24 +328,35 @@ export class FixedNumber implements ExactNumberType {
     return new FixedNumber(newNumber, newDecimalPos);
   }
 
+  private countDigits() {
+    if (this.number === 0n) return 1;
+    let digits = 0;
+    let x = this.number < 0 ? -this.number : this.number;
+    while (x > 0n) {
+      x /= 10n;
+      digits++;
+    }
+    return digits;
+  }
+
+  // move the number to the +-[0.1, 1) interval
+  private toSubZeroNum() {
+    const digits = this.countDigits();
+    const subZeroNum = new FixedNumber(this.number, digits);
+    const exponentDiff = digits - this.decimalPos;
+    return { subZeroNum, exponentDiff };
+  }
+
   roundToDigits(digits: number, roundingMode: RoundingMode): FixedNumber {
     if (!Number.isSafeInteger(digits) || digits < 1) {
       throw new Error('Invalid value for digits');
     }
 
-    const isNegative = this.number < 0n;
-    const absNumber = isNegative ? -this.number : this.number;
+    const { subZeroNum, exponentDiff } = this.toSubZeroNum();
+    let roundedNumber = subZeroNum.round(digits, roundingMode);
+    roundedNumber = roundedNumber._incExponent(exponentDiff);
 
-    // move the number to the [0.1, 1) interval
-    const str = absNumber.toString();
-    const numberBetweenZeroAndOne = new FixedNumber(absNumber, str.length);
-    let roundedNumber = numberBetweenZeroAndOne.round(digits, roundingMode);
-
-    const integerDigits = str.length - this.decimalPos;
-
-    roundedNumber = roundedNumber._incExponent(integerDigits);
-
-    return isNegative ? roundedNumber.neg() : roundedNumber;
+    return roundedNumber;
   }
 
   intPart(): ExactNumberType {
@@ -562,8 +582,9 @@ export class FixedNumber implements ExactNumberType {
   toFixed(decimals: number, roundingMode = RoundingMode.TO_ZERO): string {
     if (!Number.isSafeInteger(decimals) || decimals < 0) throw new Error('Invalid parameter');
 
-    const rounded = this.round(decimals, roundingMode);
-    return bigIntToStr(rounded.number, decimals, false);
+    const rounded = this._round(decimals, roundingMode);
+
+    return bigIntToStr(rounded.number, rounded.decimalPos, decimals, false);
   }
 
   toExponential(digits: number, roundingMode = RoundingMode.TO_ZERO): string {
@@ -643,7 +664,7 @@ export class FixedNumber implements ExactNumberType {
   toString(radix?: number, maxDigits?: number): string {
     if (radix === undefined || radix === 10) {
       const num = maxDigits !== undefined ? this.trunc(maxDigits) : this;
-      return bigIntToStr(num.number, num.decimalPos, true);
+      return bigIntToStr(num.number, num.decimalPos, num.decimalPos, true);
     }
 
     return this.toBase(radix, maxDigits);
@@ -653,8 +674,28 @@ export class FixedNumber implements ExactNumberType {
     if (!Number.isSafeInteger(digits) || digits < 1) throw new Error('Invalid parameter');
 
     const rounded = this.roundToDigits(digits, roundingMode);
-    const res = bigIntToStr(rounded.number, rounded.decimalPos, false);
-    return res;
+    const { subZeroNum, exponentDiff } = rounded.toSubZeroNum();
+
+    const isNegative = subZeroNum.sign() === -1;
+    let subZeroStr = bigIntToStr(subZeroNum.number, subZeroNum.decimalPos, subZeroNum.decimalPos, false);
+    subZeroStr = subZeroStr.slice(isNegative ? 3 : 2); // '-0.' or '0.'
+
+    // cut extra digits
+    subZeroStr = subZeroStr.slice(0, Math.max(digits, exponentDiff));
+
+    const whole = subZeroStr.slice(0, Math.max(0, exponentDiff));
+    const frac = subZeroStr.slice(Math.max(0, exponentDiff));
+
+    const suffixLength = Math.max(0, digits - whole.length - frac.length);
+    const suffix = '0'.repeat(suffixLength);
+    const prefix = '0'.repeat(exponentDiff < 0 ? -exponentDiff : 0);
+
+    let res = whole || '0';
+    if (frac.length + prefix.length + suffixLength > 0) {
+      res += `.${prefix}${frac}${suffix}`;
+    }
+
+    return isNegative ? `-${res}` : res;
   }
 
   valueOf(): number {