+++ src/Fraction.ts
@@ -1,6 +1,6 @@
 import { CommonNumberFields, ExactNumberType, ModType, RoundingMode } from './types';
 import { FixedNumber } from './FixedNumber';
-import { bigIntToStr, trimTrailingZeros } from './util';
+import { trimTrailingZeros } from './util';
 import { ExactNumber } from './ExactNumber';
 
 export class Fraction implements ExactNumberType {
@@ -39,7 +39,7 @@ export class Fraction implements ExactNumberType {
       return fraction.mul(exp).normalize() as Fraction;
     }
 
-    return fraction.normalize();
+    return fraction.simplify();
   }
 
   private parseParameter(x: number | bigint | string | ExactNumberType): Fraction {
@@ -271,11 +271,11 @@ export class Fraction implements ExactNumberType {
     }
   }
 
-  private lcm(a: bigint, b: bigint): bigint {
-    return (a * b) / this.gcd(a, b);
-  }
+  // private lcm(a: bigint, b: bigint): bigint {
+  //   return (a * b) / this.gcd(a, b);
+  // }
 
-  normalize() {
+  private simplify() {
     let { numerator, denominator } = this;
 
     const gcd = this.gcd(numerator, denominator);
@@ -292,9 +292,27 @@ export class Fraction implements ExactNumberType {
     return new Fraction(numerator, denominator);
   }
 
-  getFractionParts(normalize = true) {
-    const num = normalize ? this.normalize() : this;
+  normalize(): FixedNumber | Fraction {
+    const { numerator, denominator } = this.simplify();
 
+    if (denominator === 1n) {
+      return new FixedNumber(numerator, 0);
+    }
+
+    const frac = new Fraction(numerator, denominator);
+
+    // check if conversion to FixedNumber is possible
+    const { cycleLen, cycleStart } = frac.getDecimalFormat(0);
+
+    if (cycleLen !== 0) {
+      return frac;
+    }
+
+    return frac.round(cycleStart, RoundingMode.TO_ZERO);
+  }
+
+  getFractionParts(normalize = true) {
+    const num = normalize ? this.simplify() : this;
     return {
       numerator: new FixedNumber(num.numerator),
       denominator: new FixedNumber(num.denominator),
@@ -463,12 +481,10 @@ export class Fraction implements ExactNumberType {
     return { cycleLen, cycleStart };
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-shadow
   toFixed(decimals: number, roundingMode = RoundingMode.TO_ZERO): string {
     if (!Number.isSafeInteger(decimals) || decimals < 0) throw new Error('Invalid parameter');
 
-    const [number, decimalPos] = this.round(decimals, roundingMode).serialize();
-    return bigIntToStr(number, decimalPos, false);
+    return this.round(decimals, roundingMode).toFixed(decimals);
   }
 
   private toRepeatingParts(maxDigits: number | undefined): [string, string, string] {
@@ -476,10 +492,10 @@ export class Fraction implements ExactNumberType {
       return ['0', '', ''];
     }
 
-    const { cycleLen, cycleStart } = this.normalize().getDecimalFormat(maxDigits);
+    const { cycleLen, cycleStart } = this.simplify().getDecimalFormat(maxDigits);
 
+    // if aborted calculation or terminating decimal
     if (cycleLen === null || cycleLen === 0) {
-      // aborted calculation or terminating decimal
       const outputDigits = maxDigits ?? cycleStart;
       const str = this.toFixed(outputDigits);
       const parts = trimTrailingZeros(str).split('.');
@@ -516,7 +532,7 @@ export class Fraction implements ExactNumberType {
   }
 
   toFraction(): string {
-    const { numerator, denominator } = this.normalize();
+    const { numerator, denominator } = this.getFractionParts(true);
 
     return `${numerator.toString()}/${denominator.toString()}`;
   }
@@ -547,7 +563,7 @@ export class Fraction implements ExactNumberType {
 
     const isNegative = num.sign() === -1;
     if (isNegative) {
-      intPart = intPart.neg() as FixedNumber;
+      intPart = intPart.neg();
       fracPart = fracPart.neg();
     }
 
@@ -591,8 +607,7 @@ export class Fraction implements ExactNumberType {
   toPrecision(digits: number, roundingMode = RoundingMode.TO_ZERO): string {
     if (!Number.isSafeInteger(digits) || digits < 1) throw new Error('Invalid parameter');
 
-    const [number, decimalPos] = this.roundToDigits(digits, roundingMode).serialize();
-    return bigIntToStr(number, decimalPos, false);
+    return this.roundToDigits(digits, roundingMode).toPrecision(digits);
   }
 
   valueOf(): number {