+++ src/approx/hyperbolic.test.ts
@@ -3,6 +3,8 @@ import { compareError, testStability } from '../testHelper.test';
 
 describe('hyperbolic', () => {
   it('sinh', () => {
+    expect(sinh(0, 30).toString()).toBe('0');
+
     for (let i = -4; i <= 4; i += 0.004) {
       const jsResult = Math.sinh(i).toString();
       compareError(sinh(i.toString(), 30), jsResult);
@@ -14,6 +16,8 @@ describe('hyperbolic', () => {
   });
 
   it('cosh', () => {
+    expect(cosh(0, 30).toString()).toBe('1');
+
     for (let i = -4; i <= 4; i += 0.004) {
       const jsResult = Math.cosh(i).toString();
       compareError(cosh(i.toString(), 30), jsResult);
@@ -25,6 +29,8 @@ describe('hyperbolic', () => {
   });
 
   it('tanh', () => {
+    expect(tanh(0, 30).toString()).toBe('0');
+
     const range = [-4, 4];
     for (let i = range[0]; i <= range[1]; i += 0.004) {
       const jsResult = Math.tanh(i).toString();