+++ src/approx/hyperbolic.ts
@@ -1,6 +1,8 @@
 import { FixedNumber } from '../FixedNumber';
 import { ExactNumber } from '../ExactNumber';
 import { ExactNumberType } from '../types';
+import { sqrt } from './roots';
+import { limitDecimals } from '../util';
 
 // sinh x = x + x^3/3! + x^5/5! + ...
 function* sinhGenerator(x: ExactNumberType, decimals: number) {
@@ -20,7 +22,7 @@ function* sinhGenerator(x: ExactNumberType, decimals: number) {
 }
 
 export const sinh = (x: number | bigint | string | ExactNumberType, decimals: number): ExactNumberType => {
-  const input = ExactNumber(x);
+  const input = limitDecimals(ExactNumber(x), decimals);
 
   const maxError = new FixedNumber(`1e-${decimals + 5}`);
 
@@ -56,7 +58,7 @@ function* coshGenerator(x: ExactNumberType, decimals: number) {
 }
 
 export const cosh = (x: number | bigint | string | ExactNumberType, decimals: number): ExactNumberType => {
-  const input = ExactNumber(x);
+  const input = limitDecimals(ExactNumber(x), decimals);
 
   const maxError = new FixedNumber(`1e-${decimals + 5}`);
 
@@ -71,10 +73,15 @@ export const cosh = (x: number | bigint | string | ExactNumberType, decimals: nu
 };
 
 export const tanh = (angle: number | bigint | string | ExactNumberType, decimals: number): ExactNumberType => {
-  const angleNum = ExactNumber(angle);
+  const angleNum = limitDecimals(ExactNumber(angle), decimals);
   if (angleNum.isZero()) return ExactNumber(0);
 
   // tanh x = sinh x / cosh x;
-  const res = ExactNumber(sinh(angle, decimals + 10)).div(cosh(angle, decimals + 10));
+  // sinh x = sqrt((cosh(x)^2) - 1);
+
+  const coshRes = cosh(angleNum, decimals + 10).abs();
+  const sinhRes = sqrt(coshRes.pow(2).sub(1), decimals + 10);
+
+  const res = sinhRes.div(coshRes).mul(angleNum.sign());
   return res.trunc(decimals);
 };