+++ src/approx/roots.ts
@@ -2,6 +2,7 @@ import { FixedNumber } from '../FixedNumber';
 import { ExactNumberType } from '../types';
 import { Fraction } from '../Fraction';
 import { ExactNumber } from '../ExactNumber';
+import { limitDecimals } from '../util';
 
 const approximateNthRoot = (n: number, x: ExactNumberType): string => {
   let xNum = x.toNumber();
@@ -62,7 +63,7 @@ export const nthroot = (
   if (n < 0) throw new Error('Negative N is not supported');
   if (n === 0) throw new Error('N cannot be zero');
 
-  const xNum = ExactNumber(x);
+  const xNum = limitDecimals(ExactNumber(x), decimals);
   if (n === 1) {
     return xNum.trunc(decimals);
   }